[#15] refactor : 클래스 구조 변경에 따른 테스트 코드 변경
